abline {graphics}	R Documentation

Add a Straight Line to a Plot

Description

This function adds one or more straight lines through the current plot.

Usage

abline(a, b, untf = FALSE, ...)
abline(h=, untf = FALSE, ...)
abline(v=, untf = FALSE, ...)
abline(coef=, untf = FALSE, ...)
abline(reg=, untf = FALSE, ...)

Arguments

a,b 	 the intercept and slope.

untf 	 logical asking to untransform. See Details.

h 	 the y-value for a horizontal line.

v 	 the x-value for a vertical line.

coef 	 a vector of length two giving the intercept and slope.

reg 	 an object with a coef component. See Details.

... 	 graphical parameters.

Details

The first form specifies the line in intercept/slope form (alternatively a can be specified on its own and is taken to contain the slope and intercept in vector form).

The h= and v= forms draw horizontal and vertical lines at the specified coordinates.

The coef form specifies the line by a vector containing the slope and intercept.

reg is a regression object which contains reg$coef. If it is of length 1 then the value is taken to be the slope of a line through the origin, otherwise, the first 2 values are taken to be the intercept and slope.

If untf is true, and one or both axes are log-transformed, then a curve is drawn corresponding to a line in original coordinates, otherwise a line is drawn in the transformed coordinate system. The h and v parameters always refer to original coordinates.

The graphical parameters col and lty can be specified as arguments to abline; see par for details.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

See Also

lines and segments for connected and arbitrary lines given by their endpoints. par.

Examples

z <- lm(dist ~ speed, data = cars)
plot(cars)
abline(z)

[Package graphics version 2.1.0 Index]