graphics-defunct {graphics}	R Documentation

Defunct Functions in Package graphics

Description

The functions or variables listed here are no longer part of R as they are not needed (any more).

Usage


# Defunct in 1.3.0
dotplot(...)
stripplot(...)

# Defunct in 1.6.0
piechart(x, labels = names(x), edges = 200, radius = 0.8,
         density = NULL, angle = 45, col = NULL, main = NULL, ...)

Details

dotplot and stripplot have been renamed to dotchart and stripchart, respectively. The names are now used in lattice.

piechart was the old name for pie, but clashed with usage in Trellis (not implemented in lattice).

See Also

Defunct

[Package graphics version 3.6.1 Index]