panel.smooth {graphics}	R Documentation

Simple Panel Plot

Description

An example of a simple useful panel function to be used as argument in e.g., coplot or pairs.

Usage


panel.smooth(x, y, col = par("col"), bg = NA, pch = par("pch"),
             cex = 1, col.smooth = "red", span = 2/3, iter = 3,
             ...)

Arguments

x, y
numeric vectors of the same length

col, bg, pch, cex
numeric or character codes for the color(s), point type and size of points; see also par.

col.smooth
color to be used by lines for drawing the smooths.

span
smoothing parameter f for lowess, see there.

iter
number of robustness iterations for lowess.

...
further arguments to lines.

See Also

coplot and pairs where panel.smooth is typically used; lowess which does the smoothing.

Examples


pairs(swiss, panel = panel.smooth, pch = ".")  # emphasize the smooths
pairs(swiss, panel = panel.smooth, lwd = 2, cex = 1.5, col = "blue")  # hmm...

[Package graphics version 3.6.1 Index]