plot.window {graphics}	R Documentation

Set up World Coordinates for Graphics Window

Description

This function sets up the world coordinate system for a graphics window. It is called by higher level functions such as plot.default (after plot.new).

Usage

plot.window(xlim, ylim, log = "", asp = NA, ...)

Arguments

xlim, ylim 	 numeric of length 2, giving the x and y coordinates ranges.

log 	 character; indicating which axes should be in log scale.

asp 	 numeric, giving the aspect ratio y/x.

... 	 further graphical parameters as in par.

Details

Note that if asp is a finite positive value then the window is set up so that one data unit in the x direction is equal in length to asp * one data unit in the y direction.

The special case asp == 1 produces plots where distances between points are represented accurately on screen. Values with asp > 1 can be used to produce more accurate maps when using latitude and longitude.

The function attempts to produce a plausible set of scales if one or both of xlim and ylim is of length one or the two values given are identical, but it is better to avoid that case.

Usually, one should rather use the higher level functions such as plot, hist, image, ..., instead and refer to their help pages for explanation of the arguments.

See Also

xy.coords, plot.xy, plot.default.

Examples

##--- An example for the use of 'asp' :
require(stats)  # normally loaded
loc <- cmdscale(eurodist)
rx <- range(x <- loc[,1])
ry <- range(y <- -loc[,2])
plot(x, y, type="n", asp=1, xlab="", ylab="")
abline(h = pretty(rx, 10), v = pretty(ry, 10), col = "lightgray")
text(x, y, names(eurodist), cex=0.8)

[Package graphics version 2.1.0 Index]