plot.xy {graphics}	R Documentation

Basic Internal Plot Function

Description

This is the internal function that does the basic plotting of points and lines. Usually, one should rather use the higher level functions instead and refer to their help pages for explanation of the arguments.

Usage

plot.xy(xy, type, pch = 1, lty = "solid", col = par("fg"), bg = NA,
        cex = 1, lwd = par("lwd"), ...)

Arguments

xy 	 A four-element list as results from xy.coords.

type 	 1 character code: see plot.default.

pch 	 character or integer code for kind of points/lines, see points.default.

lty 	 line type code, see lines.

col 	 color code or name, see colors, palette.

bg 	 background (“fill”) color for the open plot symbols 21:25: see points.default.

cex 	 character expansion.

lwd 	 line width, also used for (non-filled) plot symbols, see lines and points.

... 	 further graphical parameters.

Details

The arguments pch, col, bg, cex, lwd may be vectors and may be recycled, depending on type: see points and lines for specifics.

See Also

plot, plot.default, points, lines.

Examples

points.default # to see how it calls "plot.xy(xy.coords(x, y), ...)"

[Package graphics version 2.1.0 Index]