plot.formula {graphics}	R Documentation

Formula Notation for Scatterplots

Description

Specify a scatterplot or add points or lines via a formula.

Usage

## S3 method for class 'formula':
plot(formula, data = parent.frame(), ..., subset,
             ylab = varnames[response], ask = dev.interactive())

## S3 method for class 'formula':
points(formula, data = parent.frame(), ..., subset)

## S3 method for class 'formula':
lines(formula, data = parent.frame(), ..., subset)

Arguments

formula 	 a formula, such as y ~ x.

data 	 a data.frame (or list) from which the variables in formula should be taken.

... 	 Arguments to be passed to or from other methods. horizontal = TRUE is also accepted.

subset 	 an optional vector specifying a subset of observations to be used in the fitting process.

ylab 	 the y label of the plot(s).

ask 	 logical, see par.

Details

Both the terms in the formula and the ... arguments are evaluated in data enclosed in parent.frame() if data is a list or a data frame. The terms of the formula and those arguments in ... that are of the same length as data are subjected to the subsetting specified in subset. If the formula in plot.formula contains more than one non-response term, a series of plots of y against each term is given. A plot against the running index can be specified as plot(y ~ 1).

Missing values are not considered in these methods, and in particular cases with missing values are not removed.

If y is an object (i.e. has a class attribute) then plot.formula looks for a plot method for that class first. Otherwise, the class of x will determine the type of the plot. For factors this will be a parallel boxplot, and argument horizontal = TRUE can be used (see boxplot).

Value

These functions are invoked for their side effect of drawing in the active graphics device.

See Also

plot.default, points, lines, plot.factor.

Examples

op <- par(mfrow=c(2,1))
plot(Ozone ~ Wind, data = airquality, pch=as.character(Month))
plot(Ozone ~ Wind, data = airquality, pch=as.character(Month),
     subset = Month != 7)
par(op)

[Package graphics version 2.9.2 Index]