graphics-package {graphics}	R Documentation

The R Graphics Package

Description

R functions for base graphics

Details

This package contains functions for base graphics. Base graphics are traditional S graphics, as opposed to the newer grid graphics.

For a complete list of functions with individual help pages, use library(help="graphics").

Author(s)

R Development Core Team and contributors worldwide

Maintainer: R Core Team R-core@r-project.org

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Murrell, P. (2005) R Graphics. Chapman & Hall/CRC Press.

[Package graphics version 2.9.2 Index]