hist.POSIXt {graphics}	R Documentation

Histogram of a Date or Date-Time Object

Description

Method for hist applied to date or date-time objects.

Usage

## S3 method for class 'POSIXt':
hist(x, breaks, ...,
     xlab = deparse(substitute(x)),
     plot = TRUE, freq = FALSE,
     start.on.monday = TRUE, format)

## S3 method for class 'Date':
hist(x, breaks, ...,
     xlab = deparse(substitute(x)),
     plot = TRUE, freq = FALSE,
     start.on.monday = TRUE, format)

Arguments

x 	 an object inheriting from class "POSIXt" or "Date".

breaks 	 a vector of cut points or number giving the number of intervals which x is to be cut into or an interval specification, one of "days", "weeks", "months", "quarters" or "years", plus "secs", "mins", "hours" for date-time objects.

... 	 graphical parameters, or arguments to hist.default such as include.lowest, right and labels.

xlab 	 a character string giving the label for the x axis, if plotted.

plot 	 logical. If TRUE (default), a histogram is plotted, otherwise a list of breaks and counts is returned.

freq 	 logical; if TRUE, the histogram graphic is a representation of frequencies, i.e, the counts component of the result; if FALSE, relative frequencies (probabilities) are plotted.

start.on.monday 	 logical. If breaks = "weeks", should the week start on Mondays or Sundays?

format 	 for the x-axis labels. See strptime.

Details

Using breaks = "quarters" will create intervals of 3 calendar months, with the intervals beginning on January 1, April 1, July 1 or October 1, based upon min(x) as appropriate.

Value

An object of class "histogram": see hist.

See Also

seq.POSIXt, axis.POSIXct, hist

Examples

hist(.leap.seconds, "years", freq = TRUE)
hist(.leap.seconds,
     seq(ISOdate(1970, 1, 1), ISOdate(2010, 1, 1), "5 years"))

## 100 random dates in a 10-week period
random.dates <- as.Date("2001/1/1") + 70*stats::runif(100)
hist(random.dates, "weeks", format = "%d %b")

[Package graphics version 2.9.2 Index]