plot.design {graphics}	R Documentation

Plot Univariate Effects of a ‘Design’ or Model

Description

Plot univariate effects of one ore more factors, typically for a designed experiment as analyzed by aov(). Further, in S this a method of the plot generic function for design objects.

Usage

plot.design(x, y = NULL, fun = mean, data = NULL, ...,
            ylim = NULL, xlab = "Factors", ylab = NULL,
            main = NULL, ask = NULL, xaxt = par("xaxt"),
            axes = TRUE, xtick = FALSE)

Arguments

x 	 either a data frame containing the design factors and optionally the response, or a formula or terms object.

y 	 the response, if not given in x.

fun 	 a function (or name of one) to be applied to each subset. It must return one number for a numeric (vector) input.

data 	 data frame containing the variables referenced by x when that is formula like.

... 	 graphical arguments such as col, see par.

ylim 	 range of y values, as in plot.default.

xlab 	 x axis label, see title.

ylab 	 y axis label with a ‘smart’ default.

main 	 main title, see title.

ask 	 logical indicating if the user should be asked before a new page is started – in the case of multiple y's.

xaxt 	 character giving the type of x axis.

axes 	 logical indicating if axes should be drawn.

xtick 	 logical indicating if ticks (one per factor) should be drawn on the x axis.

Details

The supplied function will be called once for each level of each factor in the design and the plot will show these summary values. The levels of a particular factor are shown along a vertical line, and the overall value of fun() for the response is drawn as a horizontal line.

This is a new R implementation which will not be completely compatible to the earlier S implementations. This is not a bug but might still change.

Note

A big effort was taken to make this closely compatible to the S version. However, col (and fg) specification has different effects.

Author(s)

Roberto Frisullo and Martin Maechler

References

Chambers, J. M. and Hastie, T. J. eds (1992) Statistical Models in S. Chapman & Hall, London, the white book, pp. 546–7 (and 163–4).

Freeny, A. E. and Landwehr, J. M. (1990) Displays for data from large designed experiments; Computer Science and Statistics: Proc. 22nd Symp. Interface, 117–126, Springer Verlag.

See Also

interaction.plot for a ‘standard graphic’ of designed experiments.

Examples

require(stats)
plot.design(warpbreaks)# automatic for data frame with one numeric var.

Form <- breaks ~ wool + tension
summary(fm1 <- aov(Form, data = warpbreaks))
plot.design(       Form, data = warpbreaks, col = 2)# same as above

## More than one y :
utils::str(esoph)
plot.design(esoph) ## two plots; if interactive you are "ask"ed

## or rather, compare mean and median:
op <- par(mfcol = 1:2)
plot.design(ncases/ncontrols ~ ., data = esoph, ylim = c(0, 0.8))
plot.design(ncases/ncontrols ~ ., data = esoph, ylim = c(0, 0.8),
            fun = median)
par(op)

[Package graphics version 2.9.2 Index]