plot.table {graphics}	R Documentation

Plot Methods for ‘table’ Objects

Description

This is a method of the generic plot function for (contingency) table objects. Whereas for two- and more dimensional tables, a mosaicplot is drawn, one-dimensional ones are plotted as bars.

Usage

## S3 method for class 'table':
plot(x, type = "h", ylim = c(0, max(x)), lwd = 2,
     xlab = NULL, ylab = NULL, frame.plot = is.num, ...)

Arguments

x 	 a table (like) object.

type 	 plotting type.

ylim 	 range of y-axis.

lwd 	 line width for bars when type = "h" is used in the 1D case.

xlab, ylab 	 x- and y-axis labels.

frame.plot 	 logical indicating if a frame (box) should be drawn in the 1D case. Defaults to true when x has dimnames coerce-able to numbers.

... 	 further graphical arguments, see plot.default.

See Also

plot.factor, the plot method for factors.

Examples

## 1-d tables
(Poiss.tab <- table(N = stats::rpois(200, lambda = 5)))
plot(Poiss.tab, main = "plot(table(rpois(200, lambda = 5)))")

plot(table(state.division))

## 4-D :
plot(Titanic, main ="plot(Titanic, main= *)")

[Package graphics version 2.9.2 Index]