plot.xy {graphics}	R Documentation

Basic Internal Plot Function

Description

This is the internal function that does the basic plotting of points and lines. Usually, one should rather use the higher level functions instead and refer to their help pages for explanation of the arguments.

Usage

plot.xy(xy, type, pch = par("pch"), lty = par("lty"),
        col = par("col"), bg = NA,
        cex = 1, lwd = par("lwd"), ...)

Arguments

xy 	 A four-element list as results from xy.coords.

type 	 1 character code: see plot.default. NULL is accepted as a synonym for "p".

pch 	 character or integer code for kind of points, see points.default.

lty 	 line type code, see lines.

col 	 color code or name, see colors, palette. Here NULL means colour 0.

bg 	 background (fill) color for the open plot symbols 21:25: see points.default.

cex 	 character expansion.

lwd 	 line width, also used for (non-filled) plot symbols, see lines and points.

... 	 further graphical parameters such as xpd, lend, ljoin and lmitre.

Details

The arguments pch, col, bg, cex, lwd may be vectors and may be recycled, depending on type: see points and lines for specifics. In particular note that lwd is treated as a vector for points and as a single (first) value for lines.

cex is a numeric factor in addition to par("cex") which affects symbols and characters as drawn by type "p", "o", "b" and "c".

See Also

plot, plot.default, points, lines.

Examples

points.default # to see how it calls "plot.xy(xy.coords(x, y), ...)"

[Package graphics version 2.9.2 Index]