plot.histogram {graphics}	R Documentation

Plot Histograms

Description

These are methods for objects of class "histogram", typically produced by hist.

Usage

## S3 method for class 'histogram':
plot(x, freq = equidist, density = NULL, angle = 45,
               col = NULL, border = par("fg"), lty = NULL,
               main = paste("Histogram of",
                            paste(x$xname, collapse="\n")),
               sub = NULL, xlab = x$xname, ylab,
               xlim = range(x$breaks), ylim = NULL,
               axes = TRUE, labels = FALSE, add = FALSE, ann = TRUE, ...)

## S3 method for class 'histogram':
lines(x, ...)

Arguments

x 	 a histogram object, or a list with components density, mid, etc, see hist for information about the components of x.

freq 	 logical; if TRUE, the histogram graphic is to present a representation of frequencies, i.e, x$counts; if FALSE, relative frequencies (probabilities), i.e., x$density, are plotted. The default is true for equidistant breaks and false otherwise.

col 	 a colour to be used to fill the bars. The default of NULL yields unfilled bars.

border 	 the color of the border around the bars.

angle, density 	 select shading of bars by lines: see rect.

lty 	 the line type used for the bars, see also lines.

main, sub, xlab, ylab 	 these arguments to title have useful defaults here.

xlim, ylim 	 the range of x and y values with sensible defaults.

axes 	 logical, indicating if axes should be drawn.

labels 	 logical or character. Additionally draw labels on top of bars, if not FALSE; if TRUE, draw the counts or rounded densities; if labels is a character, draw itself.

add 	 logical. If TRUE, only the bars are added to the current plot. This is what lines.histogram(*) does.

ann 	 logical. Should annotations (titles and axis titles) be plotted?

... 	 further graphical parameters to title and axis.

Details

lines.histogram(*) is the same as plot.histogram(*, add = TRUE).

See Also

hist, stem, density.

Examples

(wwt <- hist(women$weight, nclass = 7, plot = FALSE))
plot(wwt, labels = TRUE) # default main & xlab using wwt$xname
plot(wwt, border = "dark blue", col = "light blue",
     main = "Histogram of 15 women's weights", xlab = "weight [pounds]")

## Fake "lines" example, using non-default labels:
w2 <- wwt; w2$counts <- w2$counts - 1
lines(w2, col = "Midnight Blue", labels = ifelse(w2$counts, "> 1", "1"))

[Package graphics version 2.9.2 Index]