rug {graphics}	R Documentation

Add a Rug to a Plot

Description

Adds a rug representation (1-d plot) of the data to the plot.

Usage

rug(x, ticksize = 0.03, side = 1, lwd = 0.5, col = par("fg"),
    quiet = getOption("warn") < 0, ...)

Arguments

x 	 A numeric vector

ticksize 	 The length of the ticks making up the ‘rug’. Positive lengths give inwards ticks.

side 	 On which side of the plot box the rug will be plotted. Normally 1 (bottom) or 3 (top).

lwd 	 The line width of the ticks. Some devices will round the default width up to 1.

col 	 The colour the ticks are plotted in.

quiet 	 logical indicating if there should be a warning about clipped values.

... 	 further arguments, passed to axis, such as line or pos for specifying the location of the rug.

Details

Because of the way rug is implemented, only values of x that fall within the plot region are included. There will be a warning if any finite values are omitted, but non-finite values are omitted silently.

Prior to R 2.8.0 rug re-drew the axis like: it no longer does so.

References

Chambers, J. M. and Hastie, T. J. (1992) Statistical Models in S. Wadsworth & Brooks/Cole.

See Also

jitter which you may want for ties in x.

Examples

require(stats)# both 'density' and its default method
with(faithful, {
    plot(density(eruptions, bw = 0.15))
    rug(eruptions)
    rug(jitter(eruptions, amount = 0.01), side = 3, col = "light blue")
})

[Package graphics version 2.9.2 Index]