stem {graphics}	R Documentation

Stem-and-Leaf Plots

Description

stem produces a stem-and-leaf plot of the values in x. The parameter scale can be used to expand the scale of the plot. A value of scale=2 will cause the plot to be roughly twice as long as the default.

Usage

stem(x, scale = 1, width = 80, atom = 1e-08)

Arguments

x 	 a numeric vector.

scale 	 This controls the plot length.

width 	 The desired width of plot.

atom 	 a tolerance.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Examples

stem(islands)
stem(log10(islands))

[Package graphics version 2.9.2 Index]